+++ submission/submission-webapp/src/itest/kotlin/ac/uk/ebi/biostd/itest/common/TestUserDataService.kt
@@ -1,12 +1,11 @@
-package ac.uk.ebi.biostd.data.service
+package ac.uk.ebi.biostd.itest.common
 
 import ac.uk.ebi.biostd.persistence.model.DbUserData
 import ac.uk.ebi.biostd.persistence.repositories.UserDataDataRepository
 
-class UserDataService(
+class TestUserDataService(
     private val dataRepository: UserDataDataRepository,
 ) {
-
     fun getUserData(userEmail: String, key: String): DbUserData? = dataRepository.findByUserEmailAndKey(userEmail, key)
     fun delete(userEmail: String, accNo: String) = dataRepository.deleteByUserEmailAndKey(userEmail, accNo)
 }
